refactor(server): split worker handle & tasks for CE and EE  support
