+++ packages/server/src/ee/worker/handler/posts.ts
@@ -1,0 +1,17 @@
+import * as postsWorkerService from "../tasks/posts";
+
+import { checkLicense as checkLicenseService } from "../../do-not-remove/services/checkLicense";
+
+export async function getPostsWorker() {
+  const result = await checkLicenseService();
+
+  if ("code" in result || result.status !== "active") {
+    return {
+      posts: postsWorkerService,
+    };
+  }
+
+  return {
+    posts: {},
+  };
+}
