+++ packages/server/src/worker/tasks/index.ts
@@ -1,4 +1,12 @@
 import { mailQueue } from "./mail";
 import { postsQueue } from "./posts";
 
-export default [mailQueue, postsQueue];
+let queues = [mailQueue, postsQueue];
+
+(async () => {
+  try {
+    queues = await (await import("../../ee/worker/tasks")).getEEQueues();
+  } catch (_) {}
+})();
+
+export default queues;
