+++ packages/server/src/worker/tasks/mail.ts
@@ -1,13 +1,12 @@
 import { type ConnectionOptions, Queue, type QueueOptions } from "bullmq";
 
 import type { IPasswordResetJwtPayload } from "../../types";
-import type { ISendPostRoadmapChangeMailPayload } from "../../services/mail/types";
 
 type BaseQueueOptions = Omit<QueueOptions, "connection">;
 
 export const mailQueueName = "trx-mails";
-class MailQueue {
-  private queueInstance: Queue;
+export class MailQueue {
+  queueInstance: Queue;
 
   private readonly queueName: string;
   private readonly queueOptions: BaseQueueOptions;
@@ -30,21 +29,6 @@ class MailQueue {
 
   sendPasswordResetTokenMail(payload: IPasswordResetJwtPayload) {
     return this.queueInstance.add("sendPasswordResetTokenMail", payload);
-  }
-
-  /**
-   * Send post roadmap change mail accepts an array of ISendPostRoadmapChangeMailPayload.
-   * Since there can be more 10/10K/1M voters on a post.
-   * @param payloads
-   */
-  sendPostRoadmapChangeMail(
-    payloads: Array<ISendPostRoadmapChangeMailPayload>,
-  ) {
-    const jobs = payloads.map((payload) => ({
-      name: "sendPostRoadmapChangeMail",
-      data: payload,
-    }));
-    return this.queueInstance.addBulk(jobs);
   }
 }
